# myrpc/myrpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "",
    "title": ""
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/people": {
      "post": {
        "summary": "People.CreatePerson",
        "operationId": "CreatePerson",
        "responses": {
          "default": {
            "description": "Description",
            "schema": {
              "$ref": "#/definitions/myrpcCreatePersonResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/myrpcCreatePersonRequest"
            }
          }
        ],
        "tags": [
          "People"
        ]
      }
    },
    "/api/v1/people/{name}": {
      "get": {
        "summary": "People.GetPerson",
        "operationId": "GetPerson",
        "responses": {
          "default": {
            "description": "Description",
            "schema": {
              "$ref": "#/definitions/myrpcGetPersonResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "string"
          }
        ],
        "tags": [
          "People"
        ]
      }
    }
  },
  "definitions": {
    "myrpcCreatePersonRequest": {
      "properties": {
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "myrpcCreatePersonResponse": {
      "properties": {
        "result": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "myrpcEmptyMessage": {},
    "myrpcGetPersonRequest": {
      "properties": {
        "name": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "myrpcGetPersonResponse": {
      "properties": {
        "age": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string",
          "format": "string"
        }
      }
    }
  }
}